@@ -217,13 +217,10 @@ class FancyArray extends \ArrayObject
         return join($glue, $this->to_a());
     }
 
-    /**
-     * @return FancyArray
-     */
-    public function ksort() {
+    public function ksort(int $flags = SORT_REGULAR): bool {
         $arr = $this->to_a();
         ksort($arr);
-        return new self($arr);
+        return (bool) new self($arr);
     }
 
     /**
@@ -235,22 +232,16 @@ class FancyArray extends \ArrayObject
         return new self($arr);
     }
 
-    /**
-     * @return FancyArray
-     */
-    public function uksort($func) {
+    public function uksort(callable $callback): bool {
         $arr = $this->to_a();
-        uksort($arr, $func);
-        return new self($arr);
+        uksort($arr, $callback);
+        return (bool) new self($arr);
     }
 
-    /**
-     * @return FancyArray
-     */
-    public function asort() {
+    public function asort(int $flags = SORT_REGULAR): bool {
         $arr = $this->to_a();
         asort($arr);
-        return new self($arr);
+        return (bool) new self($arr);
     }
 
     /**
@@ -271,13 +262,10 @@ class FancyArray extends \ArrayObject
         return new self($arr);
     }
 
-    /**
-     * @return FancyArray
-     */
-    public function uasort($func) {
+    public function uasort(callable $callback): bool {
         $arr = $this->to_a();
-        uasort($arr, $func);
-        return new self($arr);
+        uasort($arr, $callback);
+        return (bool) new self($arr);
     }
 
     public function key_by($rekey_by) {
