@@ -27,9 +27,9 @@ class FilteredSequence extends FilterIterator {
      * Necessary to support FilterIterator - true = keep, false = skip
      * @return bool
      */
-    public function accept() {
+    public function accept(): bool {
         /** @var Closure $fn */
         $fn = $this->fnFilterFunction;
-        return $fn($this->current(), $this->key());
+        return $fn($this->current(), $this->key()) ?? false;
     }
 }
