@@ -37,12 +37,12 @@ class MappedSequence extends Sequence {
         $this->fnMapKeyFunction = $fnMapKeyFunction;
     }
 
-    public function current() {
+    public function current(): mixed {
         $fn = $this->fnMapValueFunction;
         return $fn(parent::current(), parent::key());
     }
 
-    public function key() {
+    public function key(): mixed {
         $fn = $this->fnMapKeyFunction;
         return $fn(parent::key(), parent::current());
     }
