@@ -10,10 +10,7 @@ class RecursiveSequence extends Sequence implements RecursiveIterator  {
         return ($this->depth - 1) !== 0;
     }
 
-    /**
-     * @return RecursiveSequence
-     */
-    public function getChildren() {
+    public function getChildren(): ?RecursiveIterator {
         $x = $this->current();
         if ($this->canGoDeeper()) {
             return self::make($x)->setMaxDepth($this->depth - 1);
@@ -34,7 +31,7 @@ class RecursiveSequence extends Sequence implements RecursiveIterator  {
     /**
      * @return bool - true if we can make a sequence out of the current item.
      */
-    public function hasChildren() {
+    public function hasChildren(): bool {
         return $this->valid() && $this->depth != 0 && self::canBeSequence($this->current());
     }
 }
