@@ -2,7 +2,9 @@
 
 namespace Revinate\Sequence;
 
-class TraverseSequence extends RecursiveSequence {
+use RecursiveIterator;
+
+class TraverseSequence extends RecursiveSequence implements RecursiveIterator {
     protected $path;
     protected $pathSeparator = '.';
 
@@ -20,10 +22,7 @@ class TraverseSequence extends RecursiveSequence {
         return $traverseSequence;
     }
 
-    /**
-     * @return TraverseSequence|MappedSequence
-     */
-    public function getChildren() {
+    public function getChildren(): ?RecursiveIterator {
         $x = $this->current();
         if ($this->canGoDeeper()) {
             return self::make($x, $this->key(), $this->pathSeparator)->setMaxDepth($this->depth - 1);
@@ -36,10 +35,7 @@ class TraverseSequence extends RecursiveSequence {
         }
     }
 
-    /**
-     * @return string
-     */
-    public function key() {
+    public function key(): string {
         $key = parent::key();
         return $this->path . $key;
     }
